git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp-misc/trunk/netperfmeter@337 acdd8558-d4fd-0310-b0a5-8cf3c1ec824a

+++ src/plotter.R
@@ -1620,8 +1620,10 @@ analyseCounterResults <- function(data, lowerLimit, upperLimit,
 processPDFbyGhostscript <- function(file)
 {
    tempFile <- paste(sep="", tempfile("embedFontsIntoPDF"), ".eps")
-   cmd1 <- paste(sep="", "pdf2ps -dPDFSETTINGS=/prepress ", file, " ", tempFile)
-   cmd2 <- paste(sep="", "ps2pdf -dPDFSETTINGS=/prepress ", tempFile, " ", file)
+   cmd1 <- paste(sep="", "pdf2ps -dPDFSETTINGS=/prepress ", file, " ", tempFile,
+                         " || mv ", file, " ", tempFile)
+   cmd2 <- paste(sep="", "ps2pdf -dPDFSETTINGS=/prepress ", tempFile, " ", file,
+                         " || mv ", tempFile, " ", file)
    ret1 <- system(cmd1)
    if(ret1 != 0) {
       stop(gettextf("status %d in running command '%s'", ret1, cmd1))
@@ -1630,5 +1632,7 @@ processPDFbyGhostscript <- function(file)
    if(ret2 != 0) {
       stop(gettextf("status %d in running command '%s'", ret2, cmd2))
    }
-   file.remove(tempFile)
+   if(file.exists(tempFile)) {
+      file.remove(tempFile)
+   }
 }
